extracted buisness logic from the AddReview box component
